+++ src/apps/seelenweg-hitbox/index.tsx
@@ -1,4 +1,5 @@
 import { wrapConsole } from '../utils/ConsoleWrapper';
+import { invoke } from '@tauri-apps/api/core';
 import { PhysicalPosition, PhysicalSize } from '@tauri-apps/api/dpi';
 import { emitTo } from '@tauri-apps/api/event';
 import { getCurrent } from '@tauri-apps/api/webviewWindow';
@@ -11,9 +12,32 @@ async function Main() {
 
   view.listen('init', () => {
     getCurrent().show();
+
     document.body.addEventListener('mousemove', () => {
       emitTo('seelenweg', 'mouseenter');
     });
+
+    async function onClick(e: MouseEvent | TouchEvent) {
+      invoke('ensure_hitboxes_zorder');
+
+      let x = 0;
+      let y = 0;
+      if (e instanceof MouseEvent) {
+        x = e.clientX;
+        y = e.clientY;
+        emitTo('fancy-toolbar', 'click', { x, y });
+        return;
+      }
+
+      if (e.touches && e.touches.length > 0) {
+        x = e.touches[0]?.clientX || 0;
+        y = e.touches[0]?.clientY || 0;
+        emitTo('fancy-toolbar', 'click', { x, y });
+      }
+    }
+
+    document.body.addEventListener('click', onClick);
+    document.body.addEventListener('touchend', onClick);
   });
 
   view.listen('resize', (event) => {