+++ src/background/exposed.rs
@@ -100,6 +100,20 @@ fn is_dev_mode() -> bool {
 }
 
 #[command]
+fn ensure_hitboxes_zorder() {
+    std::thread::spawn(|| -> Result<()> {
+        let seelen = SEELEN.lock();
+        if let Some(toolbar) = seelen.toolbar() {
+            toolbar.ensure_hitbox_zorder()?;
+        }
+        if let Some(weg) = seelen.weg() {
+            weg.ensure_hitbox_zorder()?;
+        }
+        Ok(())
+    });
+}
+
+#[command]
 fn get_accent_color() -> String {
     let mut colorization: u32 = 0;
     let mut opaque_blend = windows::Win32::Foundation::BOOL(0);
@@ -154,6 +168,7 @@ pub fn register_invoke_handler(app_builder: Builder<Wry>) -> Builder<Wry> {
         get_user_envs,
         show_app_settings,
         reload_apps_configurations,
+        ensure_hitboxes_zorder,
         // Media
         media_play_pause,
         media_next,