+++ changelog.md
@@ -4,6 +4,9 @@
 ### enhancements
 - Remove unnecessary tooltip collision on toolbar items.
 
+### fix
+- Crash on restoring app in other virtual desktop using Weg.
+
 ## [1.3.1]
 ### fix
 - disable binding monitors and monitors on apps configurations for now.